Fixed data type conversions in SSL integration

+++ net/src/openssl.c
@@ -1361,7 +1361,7 @@ static int configure_ssl(ssl_socket_t *socket,
     }
 
     ERR_clear_error();
-    SSL_CTX_set_options(socket->ctx, (long) (SSL_OP_ALL | SSL_OP_NO_SSLv2));
+    SSL_CTX_set_options(socket->ctx, SSL_OP_ALL | SSL_OP_NO_SSLv2);
     SSL_CTX_set_verify(socket->ctx, SSL_VERIFY_NONE, NULL);
 
 #ifdef WITH_OPENSSL_CUSTOM_CIPHERS
